Refactor: planet selector 리팩터링
