+++ src/app/planet/[id]/layout.tsx
@@ -1,12 +1,12 @@
 import { PropsWithChildren } from 'react';
 
 import { BottomNavigation } from '~/components/BottomNavigation';
-import { PlanetSelector } from '~/modules/PlanetSelector';
+import { PlanetSelectorTopNavigation } from '~/modules/PlanetSelectorTopNavigation';
 
 const Layout = ({ children }: PropsWithChildren) => {
   return (
-    <div className="pt-t-nav">
-      <PlanetSelector />
+    <div>
+      <PlanetSelectorTopNavigation />
       {children}
       <BottomNavigation />
     </div>
