--- src/modules/PlanetSelector/CommunityList/index.ts
@@ -1,1 +1,0 @@
-export * from './CommunityList.client';
