--- src/modules/PlanetSelector/index.ts
@@ -1,1 +1,0 @@
-export * from './PlanetSelector.client';
